backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
alembic>=1.13.1
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
APScheduler>=3.10.4
requests>=2.31.0
beautifulsoup4>=4.12.3
redis>=5.0.4
tensorflow>=2.16.1
scikit-learn>=1.4.0
joblib>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
